fix(nexus-web): frame org people management as Users, not Admins
Admin is a role privilege; the people surfaces should present users/members as the primary framing. Keep invite default as member and redirect legacy /admins paths.
