+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/organizations/[orgId]/settings/admins/page.tsx
@@ -1,1 +1,10 @@
-export { default } from './admins';
+import { redirect } from 'next/navigation';
+
+/** Legacy /settings/admins path; org people management lives under Users. */
+export default function AdminsRedirectPage({
+  params,
+}: {
+  params: { orgId: string };
+}) {
+  redirect(`/organizations/${params.orgId}/settings/users`);
+}
