+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/organizations/[orgId]/settings/lib/org-settings-route.test.ts
@@ -33,6 +33,16 @@ describe('parseOrgSettingsRoute', () => {
     });
   });
 
+  it('labels the users section (not admins)', () => {
+    expect(parseOrgSettingsRoute('/organizations/org-1/settings/users')).toEqual({
+      isOrgSettingsRoute: true,
+      orgId: 'org-1',
+      isDetail: true,
+      section: 'users',
+      sectionLabel: 'Users',
+    });
+  });
+
   it('ignores a trailing slash on the index', () => {
     expect(parseOrgSettingsRoute('/organizations/org-1/settings/')).toEqual({
       isOrgSettingsRoute: true,
